cvbuilder/requirements.txt
python-dotenv
requests
openai
Jinja2